+++ NEWS.md
@@ -28,6 +28,7 @@ dispRity v1.3.5 (2020-02-12)
 ### NEW FEATURES
 
  * *New* metric: `angles`, a dimension level 2 metric that measures the angle of the main axis of each dimension in a matrix (in slopes, angles or degrees).
+ * Completely rewritten *char.diff* function. It now uses a way faster bitwise comparison architecture and comes with different distance methods as well as modular optional arguments on how to treat various special tokens ("?", "-", "&", etc.).
 
 ### BUG FIXES
 
@@ -35,6 +36,7 @@ dispRity v1.3.5 (2020-02-12)
   * updated class evaluations throughout the package for `R` version `4.0.0`: `class(.) == *` is now `is(., *)`.
   * updated `...` argument bug PR#16223.
   * In `make.metric` the argument `...` is now ignored if any `names(...)` is `"tree"` or `"phy"`.
+  * *New* options in `char.diff`: @@@
 
 ### MINOR IMPROVEMENTS